requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.1
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==7.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
django-ipware==7.0.1
dnspython==2.8.0
feedparser==6.0.12
frozenlist==1.8.0
gnews==0.4.3
idna==3.11
ipinfo==5.4.0
multidict==6.7.1
nsepython==2.97
numpy==2.4.2
pandas==3.0.1
pillow==12.1.1
propcache==0.4.1
python-dateutil==2.9.0.post0
python-ipware==3.0.0
requests==2.32.5
scipy==1.17.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
